+++ application/src/main/java/com/svprdga/infinitescrollsample/presentation/ui/extra/Design.java
@@ -1,0 +1,45 @@
+package com.svprdga.infinitescrollsample.presentation.ui.extra;
+
+public class Design {
+
+    public static final class Icon {
+
+        public static final class Light {
+
+            public static final int ALPHA_ACTIVE_INT                        = 137;
+            public static final float ALPHA_ACTIVE_FLOAT                    = 0.54f;
+            public static final int ALPHA_INACTIVE_INT                      = 76;
+            public static final float ALPHA_INACTIVE_FLOAT                  = 0.3f;
+        }
+
+        public static final class Dark {
+
+            public static final int ALPHA_ACTIVE_INT                        = 255;
+            public static final float ALPHA_ACTIVE_FLOAT                    = 1f;
+            public static final int ALPHA_INACTIVE_INT                      = 176;
+            public static final float ALPHA_INACTIVE_FLOAT                  = 0.3f;
+        }
+
+        public static final class New {
+
+            public static final float ALPHA_ACTIVE_FLOAT                    = 0.26f;
+        }
+    }
+
+    public static final class Text {
+
+        public static final class Light {
+
+            public static final float ALPHA_PRIMARY_FLOAT                   = 1f;
+            public static final float ALPHA_SECONDARY_FLOAT                 = 0.7f;
+            public static final float ALPHA_DISABLED_FLOAT                  = 0.3f;
+        }
+
+        public static final class Dark {
+
+            public static final float ALPHA_PRIMARY_FLOAT                   = 0.87f;
+            public static final float ALPHA_SECONDARY_FLOAT                 = 0.54f;
+            public static final float ALPHA_DISABLED_FLOAT                  = 0.3f;
+        }
+    }
+}